Variable cleanup

+++ main/attitude.h
@@ -21,7 +21,7 @@ Directional calibrateGyro(Directional gyro, bool hasLaunched){//returns gyro off
   static Ewma zFilter(.002);
   static Directional lastSave;//0-5 second old data
   static Directional oldSave;//5-10 second old data
-  static uint8_t counter = 0;
+  static uint16_t counter = 0;
 
   if(hasLaunched){//once launch occurs we lock in the offsets (and return early to skip the math)
     //in order to offset the delay between launch and launch detection, we use the data from 5-10 seconds prior
@@ -64,7 +64,7 @@ Directional getAttitude(Directional gyro, bool hasLaunched){//only calibrated gy
   const int hz = 100; //number of readings per second
   static Directional oldChanges; //5-10 seconds ago
   static Directional lastChanges; //0-5 seconds ago
-  static int counter = 0;
+  static uint16_t counter = 0;
   Directional currentAttitude;
 
   counter = (counter + 1) % 500;
